refactor(simulator): Finished refactoring simulator module
BREAKING CHANGE: Simulator has been refactored to a much better version than before.

Ref: #40
