--- components/simulator/entities/object_creator.py
@@ -1,4 +1,0 @@
-class ObjectFactory:
-    @staticmethod
-    def create(**kwargs):
-        pass
