+++ docs/components.simulator.model.html
@@ -198,8 +198,8 @@
 </dd></dl>
 
 <dl class="function">
-<dt id="components.simulator.model.graphics.MakeAxesActor">
-<code class="descclassname">components.simulator.model.graphics.</code><code class="descname">MakeAxesActor</code><span class="sig-paren">(</span><em>scale</em>, <em>xyzLabels</em><span class="sig-paren">)</span><a class="headerlink" href="#components.simulator.model.graphics.MakeAxesActor" title="Permalink to this definition">¶</a></dt>
+<dt id="components.simulator.model.graphics.create_axis_widget">
+<code class="descclassname">components.simulator.model.graphics.</code><code class="descname">create_axis_widget</code><span class="sig-paren">(</span><em>scale</em>, <em>xyzLabels</em><span class="sig-paren">)</span><a class="headerlink" href="#components.simulator.model.graphics.create_axis_widget" title="Permalink to this definition">¶</a></dt>
 <dd></dd></dl>
 
 <dl class="class">
@@ -249,13 +249,13 @@
 <dd></dd></dl>
 
 <dl class="function">
-<dt id="components.simulator.model.graphics.axesCube">
-<code class="descclassname">components.simulator.model.graphics.</code><code class="descname">axesCube</code><span class="sig-paren">(</span><em>ren</em>, <em>x_bound=matrix([[-1.5</em>, <em>1.5]])</em>, <em>y_bound=matrix([[-1.5</em>, <em>1.5]])</em>, <em>z_bound=matrix([[-1.5</em>, <em>1.5]])</em><span class="sig-paren">)</span><a class="headerlink" href="#components.simulator.model.graphics.axesCube" title="Permalink to this definition">¶</a></dt>
+<dt id="components.simulator.model.graphics.axes_cube">
+<code class="descclassname">components.simulator.model.graphics.</code><code class="descname">axes_cube</code><span class="sig-paren">(</span><em>ren</em>, <em>x_bound=matrix([[-1.5</em>, <em>1.5]])</em>, <em>y_bound=matrix([[-1.5</em>, <em>1.5]])</em>, <em>z_bound=matrix([[-1.5</em>, <em>1.5]])</em><span class="sig-paren">)</span><a class="headerlink" href="#components.simulator.model.graphics.axes_cube" title="Permalink to this definition">¶</a></dt>
 <dd></dd></dl>
 
 <dl class="function">
-<dt id="components.simulator.model.graphics.axesCubeFloor">
-<code class="descclassname">components.simulator.model.graphics.</code><code class="descname">axesCubeFloor</code><span class="sig-paren">(</span><em>ren</em>, <em>x_bound=matrix([[-1.5</em>, <em>1.5]])</em>, <em>y_bound=matrix([[-1.5</em>, <em>1.5]])</em>, <em>z_bound=matrix([[-1.5</em>, <em>1.5]])</em>, <em>position=matrix([[ 0.</em>, <em>-1.5</em>, <em>0. ]])</em><span class="sig-paren">)</span><a class="headerlink" href="#components.simulator.model.graphics.axesCubeFloor" title="Permalink to this definition">¶</a></dt>
+<dt id="components.simulator.model.graphics.axes_cube_floor">
+<code class="descclassname">components.simulator.model.graphics.</code><code class="descname">axes_cube_floor</code><span class="sig-paren">(</span><em>ren</em>, <em>x_bound=matrix([[-1.5</em>, <em>1.5]])</em>, <em>y_bound=matrix([[-1.5</em>, <em>1.5]])</em>, <em>z_bound=matrix([[-1.5</em>, <em>1.5]])</em>, <em>position=matrix([[ 0.</em>, <em>-1.5</em>, <em>0. ]])</em><span class="sig-paren">)</span><a class="headerlink" href="#components.simulator.model.graphics.axes_cube_floor" title="Permalink to this definition">¶</a></dt>
 <dd></dd></dl>
 
 <dl class="function">
@@ -269,13 +269,13 @@
 <dd></dd></dl>
 
 <dl class="function">
-<dt id="components.simulator.model.graphics.circleForTrajectory">
-<code class="descclassname">components.simulator.model.graphics.</code><code class="descname">circleForTrajectory</code><span class="sig-paren">(</span><em>point</em>, <em>direction</em>, <em>index=None</em><span class="sig-paren">)</span><a class="headerlink" href="#components.simulator.model.graphics.circleForTrajectory" title="Permalink to this definition">¶</a></dt>
+<dt id="components.simulator.model.graphics.circle_for_trajectory">
+<code class="descclassname">components.simulator.model.graphics.</code><code class="descname">circle_for_trajectory</code><span class="sig-paren">(</span><em>point</em>, <em>direction</em>, <em>index=None</em><span class="sig-paren">)</span><a class="headerlink" href="#components.simulator.model.graphics.circle_for_trajectory" title="Permalink to this definition">¶</a></dt>
 <dd></dd></dl>
 
 <dl class="function">
-<dt id="components.simulator.model.graphics.cubeForPath">
-<code class="descclassname">components.simulator.model.graphics.</code><code class="descname">cubeForPath</code><span class="sig-paren">(</span><em>point</em><span class="sig-paren">)</span><a class="headerlink" href="#components.simulator.model.graphics.cubeForPath" title="Permalink to this definition">¶</a></dt>
+<dt id="components.simulator.model.graphics.cube_for_path">
+<code class="descclassname">components.simulator.model.graphics.</code><code class="descname">cube_for_path</code><span class="sig-paren">(</span><em>point</em><span class="sig-paren">)</span><a class="headerlink" href="#components.simulator.model.graphics.cube_for_path" title="Permalink to this definition">¶</a></dt>
 <dd></dd></dl>
 
 <dl class="function">
