+++ docs/genindex.html
@@ -273,9 +273,9 @@
 </li>
       <li><a href="components.simulator.model.html#components.simulator.model.graphics.axesActor2d">axesActor2d() (in module components.simulator.model.graphics)</a>
 </li>
-      <li><a href="components.simulator.model.html#components.simulator.model.graphics.axesCube">axesCube() (in module components.simulator.model.graphics)</a>
+      <li><a href="components.simulator.model.html#components.simulator.model.graphics.axes_cube">axes_cube() (in module components.simulator.model.graphics)</a>
 </li>
-      <li><a href="components.simulator.model.html#components.simulator.model.graphics.axesCubeFloor">axesCubeFloor() (in module components.simulator.model.graphics)</a>
+      <li><a href="components.simulator.model.html#components.simulator.model.graphics.axes_cube_floor">axes_cube_floor() (in module components.simulator.model.graphics)</a>
 </li>
       <li><a href="components.simulator.html#components.simulator.robot_trajectory_serial_demo.axesUniversal">axesUniversal() (in module components.simulator.robot_trajectory_serial_demo)</a>
 
@@ -363,7 +363,7 @@
         <li><a href="components.simulator.common.html#components.simulator.common.pose.SO3.check">(components.simulator.common.pose.SO3 static method)</a>
 </li>
       </ul></li>
-      <li><a href="components.simulator.model.html#components.simulator.model.graphics.circleForTrajectory">circleForTrajectory() (in module components.simulator.model.graphics)</a>
+      <li><a href="components.simulator.model.html#components.simulator.model.graphics.circle_for_trajectory">circle_for_trajectory() (in module components.simulator.model.graphics)</a>
 </li>
       <li><a href="components.structure.pathplanning.searches.html#components.structure.pathplanning.searches.wavefront.Layer.columnize">columnize() (components.structure.pathplanning.searches.wavefront.Layer method)</a>
 </li>
@@ -577,7 +577,7 @@
         <li><a href="components.simulator.common.html#components.simulator.common.util.ctraj">(in module components.simulator.common.util)</a>
 </li>
       </ul></li>
-      <li><a href="components.simulator.model.html#components.simulator.model.graphics.cubeForPath">cubeForPath() (in module components.simulator.model.graphics)</a>
+      <li><a href="components.simulator.model.html#components.simulator.model.graphics.cube_for_path">cube_for_path() (in module components.simulator.model.graphics)</a>
 </li>
   </ul></td>
 </tr></table>
@@ -1169,7 +1169,7 @@
         <li><a href="components.robot.behaviors.html#components.robot.behaviors.wait.robot_trajectory_serial_demo">(in module components.robot.behaviors.wait)</a>
 </li>
       </ul></li>
-      <li><a href="components.simulator.model.html#components.simulator.model.graphics.MakeAxesActor">MakeAxesActor() (in module components.simulator.model.graphics)</a>
+      <li><a href="components.simulator.model.html#components.simulator.model.graphics.create_axis_widget">create_axis_widget() (in module components.simulator.model.graphics)</a>
 </li>
       <li><a href="components.robot.common.html#components.robot.common.super_pose.SuperPose.mat">mat (components.robot.common.super_pose.SuperPose attribute)</a>
